+++ src/main/java/io/github/opencubicchunks/cubicchunks/chunk/ChunkCubeGetter.java
@@ -1,0 +1,6 @@
+package io.github.opencubicchunks.cubicchunks.chunk;
+
+public interface ChunkCubeGetter {
+
+    IBigCube getCube(int y);
+}
